fix: disable canceling sync operation
Currently task termination is a hard-kill, which causes problems when
done on a sync operation since it might leave dangling git locks or
a partially applied `git am`.

It causes more problems than it's worth, so disable cancel for now. If
a graceful kill can be implemented (Windows is problematic) using
'SIGINT' then it could be re-enabled.